cleaning up main
